Updated Quiz Page Functionality
Added in capacity to support short text questions/answers and long paragraph questions/answers. CSS Styling not final and subject to change. Added jQuery and autoResize plugin JS files to the JS folder. 
There exists a bug to be fixed where clicking on the <label> divider of the textarea element (Question 6) goes to the top of page which is unfixed.