Installing Electron-Firebase (NOT WORKING)
This is currently not working, throwing error. Probably about NPM and the package.json. The app now has a different startup, which will be fixed and moved back. Currently working on this branch. If it is a deadend, it'll be rolled back to previous branch AUTHENTICATION and built from scratch.