Update Quiz.svelte
Changed one line to include a commit to mention that the removed node modules commit removed them from being tracked and also included a lot of changes to the svelte build which included adding the quiz page.

The notes for that commit are as follows:
The route (/quiz) was added for the quiz page and linked to the big button in the middle. The other buttons were set up but the actual routes werent made.
A dropdown was made for selecting the subject for quiz generation and a generate function which converts the subject name to the shortform name.